Added the ExploringEntryPage to the phone app.
TODO:
Fix FacebookIntegration.cs
It is not possible to log into fb on the phone, due to deprecated method being used.

I added code to change that when the phone app runs. However, it did not fix the problem.
I have asked larus to look at it when he has the time.