Changed fields to ObservableCollections, now Username is automatically shown at log-in but the profilepicture still needs the updateuser button to be pressed.
I've changed ExploringEntryPage so it's just the standart resaurant browser we all know, next up is making it different and unique because right now this sucks.

The code in the FBIntegration class is also very shitty set up, sorry about that but most of the stuff I did for the restaurants is just thrown at the bottom of the class.